+++ Changelog.md
@@ -8,6 +8,8 @@ Changes:
 
 Fixes:
 
+- Add test and comparison for invalid constraints.
+
 ## 0.1.13 - 11/3/2020
 
 Fixes: